Fix: Refactor logic for display dashboard title on navbar
